Week7/support_center/requirements.txt
mcp[cli]>=1.2.0
openai>=1.12.0
python-dotenv>=1.0.0